Simple split memory block in half

+++ MachoLoader/Classes/MemoryBlockStore.h
@@ -21,7 +21,7 @@
 
 - (SHMemoryBlock *)memoryBlockAtIndex:(int)ind;
 
-- (int)itemCount;
+- (NSUInteger)itemCount;
 - (char *)startAddress;
 - (char *)lastAddress;
